+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.2] - 2017-08-27
 ### Added
 - Add scss helper classes about text. #32
 - Add scss helper classes about display. #35
@@ -77,7 +79,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [README.md](README.md)
 - [docs/README.md](docs/README.md)
 
-[Unreleased]: https://github.com/archco/moss-ui/compare/v0.2.1...HEAD
+[Unreleased]: https://github.com/archco/moss-ui/compare/v0.2.2...HEAD
+[0.2.2]: https://github.com/archco/moss-ui/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/archco/moss-ui/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/archco/moss-ui/compare/v0.1.5...v0.2.0
 [0.1.5]: https://github.com/archco/moss-ui/compare/v0.1.4...v0.1.5